Use the signing key to define scope for a request
Previous aws-erlang and awe-elixir used the
endpoint prefix as the service name when signing
requests. This is fine for most cases, but for
services like `dynamodb streams` this breaks down,
as requests to those services should use the
`signingName` (which happens to be `dynamodb`, not
`streams.dynamodb`).

This commit fixes that, and uses the `signinName`
value from AWSs APIs to sign.